requirements.txt
tensorflow>=2.12
numpy
pillow
scikit-learn
matplotlib
fastapi
uvicorn[standard]
opencv-python
tqdm